+++ packages/forma-36-website/src/components/DocSearch.jsx
@@ -37,7 +37,7 @@ const DocSearch = () => {
 
     window.docsearch({
       // The key is added here only give access to searching the public content of the website https://docsearch.algolia.com/docs/what-is-docsearch
-      // You can even check Forma36's configuration in DocSearch's repo https://github.com/algolia/docsearch-configs/blob/master/configs/contentful_forma-36.json
+      // You can even check Forma 36's configuration in DocSearch's repo https://github.com/algolia/docsearch-configs/blob/master/configs/contentful_forma-36.json
       apiKey: 'b7d2cac8e38b0903385db259b042c66c',
       indexName: 'contentful_forma-36',
       inputSelector: '#search',